Asp .Net EnUygunHotel App
The full stack Asp.Net project with the management panel works with the MsSQL database for Crud operations, while listing the hotels in it, it lists the hotels by filtering and sorting directly in the database. With this project, it is much easier to upload images, add hotels, edit blogs, edit Master pages and corporate and about us pages.